commit5

+++ app/src/main/java/com/example/testapp/User.java
@@ -5,20 +5,20 @@ import com.google.firebase.database.IgnoreExtraProperties;
 @IgnoreExtraProperties
 public class User {
 
-    public int amountDue;
+    public long amountDue;
     public String userName ,priority ,userId;
 
     public User() {
     }
 
-    public User(int amountDue, String userName , String priority , String userId) {
+    public User(long amountDue, String userName , String priority , String userId) {
         this.amountDue = amountDue;
         this.userName = userName;
         this.priority = priority;
         this.userId =userId;
     }
 
-    public int getAmountDue() {
+    public long getAmountDue() {
         return amountDue;
     }
     public String getPriority() {
